changed file structure and uploaded makefile
The code for downloading and preparing the data for merging is now saved
in one Rsource file ("data_preperation.R") and the code for merging the
Afrobarometer rounds 3 to 5 is now saved in one Rsource file
("merging_data.R"). Both of these Rsource code files are run by a
makefile called "makefileAfrobarometer.R" which is saved in the main
folder of assignment 3.