added a word document with important variables
Some of the important variables such as country or respondent number
were still missing in the excel sheet and in these the difference in
coding is crucial as they will be the identifyer variables in the final
data set so I added them.  I also added a few other basic variables of
which I thought they could be of interest to us (age, perception of
economic condition and living conditions).