final changes
I finished the code for both the data preparation and the merging
Rsource files.
I decided to recode those variables that had values from 0 to 3 into 1
to 4, because for some reason I thought that might be better for
anlaysing the data later (maybe it was unnecessary, but it felt right).
I also tried out the makefile which links the other two Rsource files
and it also works perfectly.
So now the final dataset that we can use for analysis is stored as
"Afrobarometer.final" in the PreperatoryWork-folder